+++ changelog.md
@@ -74,6 +74,9 @@
     * `Breaker().in_transit_time`: The transition time from open to close in seconds.
 * Added `getCustomersForContainer` to `CustomerConsumerClient` which allows fetching all the `Customer`s for a given `EquipmentContainer`
 * Added `getDiagramObjects` to `DiagramConsumerClient` which allows fetching all the `DiagramObject`s matching a given mRID.
+* `Traversal` has two new helper methods:
+    * `if_not_stopping`: Adds a step action that is only called if the traversal is not stopping on the item.
+    * `if_stopping`: Adds a step action that is only called if the traversal is stopping on the item.
 
 ### Enhancements